fix issue 5813, prevent destroyAll method of DubboShutdownHook being invoke twice when using Spring

+++ dubbo-config/dubbo-config-api/src/main/java/org/apache/dubbo/config/bootstrap/DubboBootstrap.java
@@ -169,6 +169,8 @@ public class DubboBootstrap extends GenericEventListener {
 
     private List<CompletableFuture<Object>> asyncReferringFutures = new ArrayList<>();
 
+    private volatile AtomicBoolean destroyClock = new AtomicBoolean(false);
+
     /**
      * See {@link ApplicationModel} and {@link ExtensionLoader} for why DubboBootstrap is designed to be singleton.
      */
@@ -1033,23 +1035,25 @@ public class DubboBootstrap extends GenericEventListener {
     }
 
     public void destroy() {
-        // for compatibility purpose
-        DubboShutdownHook.destroyAll();
+        if (destroyClock.compareAndSet(false, true)) {
+            // for compatibility purpose
+            DubboShutdownHook.destroyAll();
 
-        if (started.compareAndSet(true, false)
-                && destroyed.compareAndSet(false, true)) {
-            unregisterServiceInstance();
-            unexportMetadataService();
-            unexportServices();
-            unreferServices();
+            if (started.compareAndSet(true, false)
+                    && destroyed.compareAndSet(false, true)) {
+                unregisterServiceInstance();
+                unexportMetadataService();
+                unexportServices();
+                unreferServices();
 
-            destroyRegistries();
-            destroyProtocols();
-            destroyServiceDiscoveries();
+                destroyRegistries();
+                destroyProtocols();
+                destroyServiceDiscoveries();
 
-            clear();
-            shutdown();
-            release();
+                clear();
+                shutdown();
+                release();
+            }
         }
     }
